Enhance startup speed when brew is installed
Check for /opt/homebrew instead of the command 'brew'. This is much
faster, but has the drawback that it will not correctly recognise a
homebrew installation on an Intel Mac (where the prefix is /usr/local).
But this is no problem, taking into consideration that zprofile also
doesn't support Intel Macs.